Fix UpdateMuzzleFlashProperties
Decouple `UpdateMuzzleFlashProperties` from BasePlayer, and move it to
CNEO_Player shared implementation.

For client side, use a `player_spawn` event listener for figuring out
when the player spawns, instead of trying to capture it in-band from a
networked "slot" command to the user. This fixes the bug that #570
attempted to address, (which will probably be reverted by #598, assuming
it's accepted), and will remove the dependency to client side weapon
switch for triggering `UpdateMuzzleFlashProperties` on player spawn.