Add cheat command "nofreeze"
Add sv_cheats protected console command "nofreeze", which allows the
player to move during a pre-round freezetime. This can be useful for
development to reduce time wasted waiting to be able to move.

Also refactor all references to the neo flags to use the
`IsInFreezetime` helper method instead of manually doing bitwise math to
check for the freezetime flag.