Refactor base ent shadow handles to 1-sized vector
This refactor has no functional changes, but is a midway commit for
preparing to support renderables with more than 1 shadow cast at
different angles.

Refactor C_BaseEntity shadow handle member variable m_ShadowHandle from
type ClientShadowHandle_t to CUtlVector<ClientShadowHandle_t> of size 1.
This commit merely changes the data type, but will still hold the
assumption that there may only be 1 shadow handle per entity.

Change GetBrightestLightSource to GetNthBrightestLightSource for
supporting getting the nth brightest light source, instead of just the
brightest. Currently, we just query it for the 0th brightest (i.e. the
brightest) source like before, but this is in preparation for additional
shadows for the other angles for the same entity.

The enum CLIENTSHADOW_INVALID_HANDLE is modified from largest value
(65535) to one less, so that we can reserve the largest value for the
new magic value CLIENTSHADOW_OUT_OF_RANGE, to be used for iterating
shadow handles, for the future >1 shadows-per-ent support. This
technically reduces the max castable shadows from 65534 to 65533, but in
practice it doesn't matter because we'll never reach this limit.

Finally, we're changing the IClientRenderable::GetShadowHandle interface
prototype to require a shadow index parameter, and refactoring any code
related to the above changes to be compatible with the new syntax.