Improve release assets compression
* Change the release archive format from ZIP to 7-Zip
* Change the compression algorithm from DEFLATE to LZMA
* Change the compression level from 6 (default) to 9 (ultra)
* Increase the GH actions/upload-artifact compression-level from 6 to 9

Optimize release assets for file size, because there should not be any
meaningful loss in compress/decompress performance for our assets data.

At time of this commit, the total release assets size before this patch
is about 946 MB. After this patch, the total size is about 638 MB. So
this change results in about 33% total size reduction for the release
assets.

Because the Windows file explorer has natively supported 7-Zip
decompression since Windows 11 22H2 (released late 2022), this change
should not affect the regular user experience for people with an
up-to-date Windows machine. And since 7z is a fairly well-known format,
I doubt it should cause confusion even for those whose machines don't
have native extraction capability.

Another alternative considered was changing the ZIP algorithm to LZMA,
as support for it was added to the ZIP spec already in 2006, but as of
this commit date, Windows still does not natively handle LZMA-compressed
ZIP archives extraction. Therefore, to avoid users confusing
LZMA-compressed ZIPs as being corrupted, I chose to instead change the
entire archive format to 7-Zip.

Note that this is a breaking change for the Steam release script, because both the
archive format (ZIP -> 7-Zip) and the file extension (".zip" -> ".7z")
have changed. The file name patterns aside from the file extension
remain unchanged.